+++ services/image/v2/model/ImageDescriptionResponseResult.ts
@@ -1,0 +1,11 @@
+
+
+export class ImageDescriptionResponseResult {
+    public description?: string;
+    public constructor() { 
+    }
+    public withDescription(description: string): ImageDescriptionResponseResult {
+        this['description'] = description;
+        return this;
+    }
+}